database/console/migrate_make_command.go
package console

import (
	"github.com/gookit/color"
	"github.com/urfave/cli/v2"

	"github.com/goravel/framework/contracts/console"
)

type MigrateMakeCommand struct {
}

//Signature The name and signature of the console command.
func (receiver *MigrateMakeCommand) Signature() string {
	return "make:migration"
}

//Description The console command description.
func (receiver *MigrateMakeCommand) Description() string {
	return "Create a new migration file"
}

//Extend The console command extend.
func (receiver *MigrateMakeCommand) Extend() console.CommandExtend {
	return console.CommandExtend{
		Category: "make",
	}
}

//Handle Execute the console command.
func (receiver *MigrateMakeCommand) Handle(c *cli.Context) error {
	// It's possible for the developer to specify the tables to modify in this
	// schema operation. The developer may also specify if this table needs
	// to be freshly created, so we can create the appropriate migrations.
	name := c.Args().First()
	if name == "" {
		color.Redln("Not enough arguments (missing: name)")

		return nil
	}

	// We will attempt to guess the table name if this the migration has
	// "create" in the name. This will allow us to provide a convenient way
	// of creating migrations that create new tables for the application.
	table, create := TableGuesser{}.Guess(name)

	//Write the migration file to disk.
	MigrateCreator{}.Create(name, table, create)

	color.Green.Printf("Created Migration: %s\n", name)

	return nil
}
